Move PulseProxy to a separate library
