+++ Package.swift
@@ -11,10 +11,12 @@ let package = Package(
     ],
     products: [
         .library(name: "Pulse", targets: ["Pulse"]),
+        .library(name: "PulseProxy", targets: ["PulseProxy"]),
         .library(name: "PulseUI", targets: ["PulseUI"])
     ],
     targets: [
         .target(name: "Pulse"),
+        .target(name: "PulseProxy", dependencies: ["Pulse"]),
         .target(name: "PulseUI", dependencies: ["Pulse"]),
     ],
     swiftLanguageVersions: [
